Move CompactionFactory to compaction-core
